Test the added logic in MenuButton and fix ReceiptList's test removed logic
