+++ src/components/__tests__/ReceiptInfo.test.js
@@ -2,8 +2,6 @@ import React from 'react'
 import { configure, shallow } from 'enzyme'
 import Adapter from 'enzyme-adapter-react-16'
 import ReceiptInfo from '../UI/ReceiptInfo/ReceiptInfo'
-
-
 configure({adapter: new Adapter()})
 
 describe('Testing <ReceiptInfo />', () => {
